Fix bug where backoff is always 0sec
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/lb/trunk@5477 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/backoffs.c
@@ -41,7 +41,7 @@ void
 xlb_backoff_server()
 {
   // DEBUG("backoff()");
-  int delay = random_between_double(0,backoff_server_max);
+  double delay = random_between_double(0,backoff_server_max);
   time_delay(delay);
 }
